refactor(gui): split app.rs into home/detail/status/widgets submodules
app.rs had grown to ~1850 lines: the Route/DetailTab/AppView state
machine plus ~35 free render helpers for two nearly disjoint screens.
Pure code motion into an app/ directory (repo foo.rs + foo/ layout):

- app/home.rs: gallery header, device cards, keyboard glow, and the
  scanning/unavailable/empty states
- app/detail.rs: detail header + tabs and the four section bodies
- app/status.rs: connecting/unreachable/outdated frames and the footer
- app/widgets.rs: shared leaf helpers (panel cards, status pills,
  header buttons, labels)

app.rs keeps Route, DetailTab, AppView, Render, and the tests.
glow_canvas/keyboard_glow stay re-exported at crate::app for
mouse_model::view. No behavior, signature, or rendering changes;
helpers are pub(super) and trait imports stay unconditional for the
non-macOS builds.
